Rename `PlayerExhaustEvent` to `EntityExhaustEvent`
