+++ src/entity/HungerManager.php
@@ -25,7 +25,7 @@ namespace pocketmine\entity;
 
 use pocketmine\event\entity\EntityDamageEvent;
 use pocketmine\event\entity\EntityRegainHealthEvent;
-use pocketmine\event\player\PlayerExhaustEvent;
+use pocketmine\event\entity\EntityExhaustEvent;
 use pocketmine\world\World;
 use function max;
 use function min;
@@ -130,13 +130,13 @@ class HungerManager{
 	 *
 	 * @return float the amount of exhaustion level increased
 	 */
-	public function exhaust(float $amount, int $cause = PlayerExhaustEvent::CAUSE_CUSTOM) : float{
+	public function exhaust(float $amount, int $cause = EntityExhaustEvent::CAUSE_CUSTOM) : float{
 		if(!$this->enabled){
 			return 0;
 		}
 		$evAmount = $amount;
-		if(PlayerExhaustEvent::hasHandlers()){
-			$ev = new PlayerExhaustEvent($this->entity, $amount, $cause);
+		if(EntityExhaustEvent::hasHandlers()){
+			$ev = new EntityExhaustEvent($this->entity, $amount, $cause);
 			$ev->call();
 			if($ev->isCancelled()){
 				return 0.0;
@@ -205,7 +205,7 @@ class HungerManager{
 			if($food >= 18){
 				if($health < $this->entity->getMaxHealth()){
 					$this->entity->heal(new EntityRegainHealthEvent($this->entity, 1, EntityRegainHealthEvent::CAUSE_SATURATION));
-					$this->exhaust(6.0, PlayerExhaustEvent::CAUSE_HEALTH_REGEN);
+					$this->exhaust(6.0, EntityExhaustEvent::CAUSE_HEALTH_REGEN);
 				}
 			}elseif($food <= 0){
 				if(($difficulty === World::DIFFICULTY_EASY && $health > 10) || ($difficulty === World::DIFFICULTY_NORMAL && $health > 1) || $difficulty === World::DIFFICULTY_HARD){
